# api-docs.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "API REST - Ocularis",
                 "description":  "Documentación de la API REST del proyecto api-rest",
                 "version":  "0.0.1-SNAPSHOT"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
                  "/api/patient/update/{id}":  {
                                                   "put":  {
                                                               "tags":  [
                                                                            "patient-controller"
                                                                        ],
                                                               "operationId":  "updatePatient",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int32"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/Patient"
                                                                                                                                       }
                                                                                                                        }
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK"
                                                                                         }
                                                                             }
                                                           }
                                               },
...
                  "/api/patient/create":  {
                                              "post":  {
                                                           "tags":  [
                                                                        "patient-controller"
                                                                    ],
                                                           "operationId":  "createPatient",
                                                           "requestBody":  {
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/Patient"
                                                                                                                                   }
                                                                                                                    }
                                                                                           },
                                                                               "required":  true
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "*/*":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/Patient"
                                                                                                                                }
                                                                                                                 }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       }
                                          },
...
                  "/api/patient/get/{id}":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "patient-controller"
                                                                     ],
                                                            "operationId":  "getPatientById",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/PatientDTO"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                            },
                  "/api/patient/address/{address}":  {
                                                         "get":  {
                                                                     "tags":  [
                                                                                  "patient-controller"
                                                                              ],
                                                                     "operationId":  "getPatientsByAddress",
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "address",
                                                                                            "in":  "path",
                                                                                            "required":  true,
                                                                                            "schema":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                        }
                                                                                    ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "OK",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "type":  "array",
                                                                                                                                              "items":  {
                                                                                                                                                            "$ref":  "#/components/schemas/PatientDTO"
                                                                                                                                                        }
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               }
                                                                                   }
                                                                 }
                                                     },
...
                  "/api/patient/delete/{id}":  {
                                                   "delete":  {
                                                                  "tags":  [
                                                                               "patient-controller"
                                                                           ],
                                                                  "operationId":  "deletePatient",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "id",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int32"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "OK"
                                                                                            }
                                                                                }
                                                              }
                                               },
...
              },
    "components":  {
                       "schemas":  {
                                       "Appointment":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "id":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int32"
                                                                                     },
                                                                              "dateTime":  {
                                                                                               "type":  "string",
                                                                                               "format":  "date"
                                                                                           },
                                                                              "patient":  {
                                                                                              "$ref":  "#/components/schemas/Patient"
                                                                                          },
                                                                              "doctor":  {
                                                                                             "$ref":  "#/components/schemas/Doctor"
                                                                                         },
                                                                              "reason":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                              "status":  {
                                                                                             "type":  "string",
                                                                                             "enum":  [
                                                                                                          "SCHEDULED",
                                                                                                          "CONFIRMED",
                                                                                                          "CANCELED",
                                                                                                          "COMPLETED",
                                                                                                          "NO_ATTEND"
                                                                                                      ]
                                                                                         }
                                                                          }
                                                       },
                                       "Doctor":  {
                                                      "type":  "object",
                                                      "properties":  {
                                                                         "id":  {
                                                                                    "type":  "integer",
                                                                                    "format":  "int32"
                                                                                },
                                                                         "dni":  {
                                                                                     "type":  "string"
                                                                                 },
                                                                         "firstName":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                         "secondName":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                         "lastName":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                         "secondLastName":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                         "email":  {
                                                                                       "type":  "string"
                                                                                   },
                                                                         "phone":  {
                                                                                       "type":  "string"
                                                                                   },
                                                                         "licenseNumber":  {
                                                                                               "type":  "string"
                                                                                           },
                                                                         "specialty":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                         "appointments":  {
                                                                                              "type":  "array",
                                                                                              "items":  {
                                                                                                            "$ref":  "#/components/schemas/Appointment"
                                                                                                        }
                                                                                          }
                                                                     }
                                                  },
                                       "Patient":  {
                                                       "type":  "object",
                                                       "properties":  {
                                                                          "id":  {
                                                                                     "type":  "integer",
                                                                                     "format":  "int32"
                                                                                 },
                                                                          "dni":  {
                                                                                      "type":  "string"
                                                                                  },
                                                                          "firstName":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                          "secondName":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                          "lastName":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                          "secondLastName":  {
                                                                                                 "type":  "string"
                                                                                             },
                                                                          "email":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                          "phone":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                          "birthDate":  {
                                                                                            "type":  "string",
                                                                                            "format":  "date"
                                                                                        },
                                                                          "address":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                          "appointments":  {
                                                                                               "type":  "array",
                                                                                               "items":  {
                                                                                                             "$ref":  "#/components/schemas/Appointment"
                                                                                                         }
                                                                                           }
                                                                      }
                                                   },
...
                                       "PatientDTO":  {
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "id":  {
                                                                                        "type":  "integer",
                                                                                        "format":  "int32"
                                                                                    },
                                                                             "dni":  {
                                                                                         "type":  "string"
                                                                                     },
                                                                             "firstName":  {
                                                                                               "type":  "string"
                                                                                           },
                                                                             "secondName":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                             "lastName":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                             "secondLastName":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                             "email":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                             "phone":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                             "birthDate":  {
                                                                                               "type":  "string",
                                                                                               "format":  "date"
                                                                                           },
                                                                             "address":  {
                                                                                             "type":  "string"
                                                                                         }
                                                                         }
                                                      },
...
                                   }
                   }
}